MAJOR CHANGES (Read description)
Updated User Permissions! There are 3 permissions: Admin, Edit, and View:

If a user is Admin, they can do all commands, Edit = all commands except for changing permissions of other users, and View = only view, no edit.

I've revamped the menu system such that we only have 4 menus now and they no longer overlap.

userMenu is the menu you see from the get go, prompting you to log in. YOU MUST LOG IN OR CREATE AN ACCOUNT TO MOVE PAST THIS MENU

Main menu is the traditional menu for everyone; this is the default menu after logging in

editmenu and viewmenu are special menus created for those with permissions, removing their access to certain functions.

I've created a new document called menus.py which simply houses all the menus that we have since they're very bulky and don't belong in functions.py.

All functionality should remain the same.

PULL THIS ASAP
