+++ users.json
@@ -1,2 +1,1 @@
-{"Dylan": {"password": "3733cd977ff8eb18b987357e22ced99f46097f31ecb239e878ae63760e83e4d5", "security_questions": {"Security Question 1: Where were you born?": "a", "Security Question 2: What is your favorite meal?": "a", "Security Question 3: What city did your parents meet?": "a"}, "email": "dtalm001@ucr.edu"}, "Brandon": {"password": "f36cf527e213d53f3b1926ee3cf99521fe61aa6d09528589b72025cee49dd177", "security_questions": {"Security Question 1: Where were you born?": "a", "Security Question 2: What is your favorite meal?": "a", "Security Question 3: What city did your parents meet?": "a"}, "email": "btalmood@gmail.com"}}
-
+{}
